refactor(harness): move the pi loader rule into internal/pluginformat
The rule that decides whether pi's `-e <dir>` loader would load anything
from a directory is not harness policy — it mirrors pi's own source, and
internal/runtime needs the same reserved names and tree-entry rule. Move
it, with its tests, into a leaf package that neither internal/harness nor
internal/runtime sits below, and give it the verdict the harness needs
next: Detect for a local directory, DetectTree for a fetched tree.

plugin.json at the directory root is checked first and settles the
verdict, so a Claude plugin that bundles a Node MCP server — whose
package.json "main" would also satisfy pi's rule — is not read as a pi
extension as well.

The only behaviour change to `extensions:` is a sharper message when the
named directory turns out to be a Claude plugin.
