+++ internal/harness/extension_spec_test.go
@@ -3,12 +3,13 @@ package harness
 import (
 	"os"
 	"path/filepath"
-	"strings"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"gopkg.in/yaml.v3"
+
+	"github.com/fullsend-ai/fullsend/internal/pluginformat"
 )
 
 func TestExtensionSpec_UnmarshalStringForm(t *testing.T) {
@@ -244,471 +245,67 @@ func TestResolveRelativeTo_Extensions(t *testing.T) {
 	assert.Contains(t, err.Error(), "extensions[0]")
 }
 
-func writeExtDir(t *testing.T, files map[string]string) string {
-	t.Helper()
-	dir := filepath.Join(t.TempDir(), "my-ext")
-	for name, content := range files {
-		p := filepath.Join(dir, name)
-		require.NoError(t, os.MkdirAll(filepath.Dir(p), 0o755))
-		require.NoError(t, os.WriteFile(p, []byte(content), 0o644))
-	}
-	require.NoError(t, os.MkdirAll(dir, 0o755))
-	return dir
-}
-
-func TestValidateFilesExist_ExtensionLoadable(t *testing.T) {
-	t.Parallel()
-	agent := filepath.Join(t.TempDir(), "code.md")
-	require.NoError(t, os.WriteFile(agent, []byte("# agent"), 0o644))
-
-	ok := map[string]map[string]string{
-		"index.js":                {"index.js": "export default function () {}"},
-		"index.ts":                {"index.ts": "export default function () {}"},
-		"index.mjs":               {"index.mjs": "export default function () {}"},
-		"index.cjs":               {"index.cjs": "module.exports = function () {}"},
-		"package.json entries":    {"package.json": `{"name":"x","pi":{"extensions":["src/main.js"]}}`, "src/main.js": "//"},
-		"package.json main":       {"package.json": `{"name":"x","main":"dist/ext.js"}`, "dist/ext.js": "//"},
-		"package.json without pi": {"package.json": `{"name":"x"}`, "index.js": "//"},
-		// pi.extensions is the explicit form and wins outright: a package
-		// resource directory does not shadow it.
-		"pi entries with skills dir": {"package.json": `{"pi":{"extensions":["index.js"]}}`, "index.js": "//", "skills/s/SKILL.md": "#"},
-		"vendored deps beside index": {"index.js": "//", "node_modules/dep/index.js": "//"},
-	}
-	for name, files := range ok {
-		t.Run("ok/"+name, func(t *testing.T) {
-			h := &Harness{Agent: agent, Extensions: []ExtensionSpec{{Path: writeExtDir(t, files)}}}
-			require.NoError(t, h.ValidateFilesExist())
-		})
-	}
-
-	// pi exits 1 with `Failed to load extension … Cannot find module` for
-	// each of these, so validation has to refuse them.
-	noEntry := map[string]map[string]string{
-		"empty":                   {},
-		"only nested js":          {"src/main.js": "//"},
-		"only README":             {"README.md": "#"},
-		"top-level js only":       {"tools.js": "//", "README.md": "#"},
-		"top-level ts only":       {"tools.ts": "//"},
-		"subdir index only":       {"sub/index.js": "//"},
-		"main missing":            {"package.json": `{"main":"dist/ext.js"}`},
-		"package.json unparsable": {"package.json": `{`},
-		"node_modules only":       {"node_modules/dep/index.js": "//"},
-	}
-	for name, files := range noEntry {
-		t.Run("no-entry/"+name, func(t *testing.T) {
-			dir := writeExtDir(t, files)
-			h := &Harness{Agent: agent, Extensions: []ExtensionSpec{{Path: dir}}}
-			err := h.ValidateFilesExist()
-			require.Error(t, err)
-			assert.Equal(t, `extensions[0] "`+dir+`": no index.js/index.ts/index.mjs/index.cjs, package.json "pi.extensions" entry or "main" file — pi would fail to load it`, err.Error())
-		})
-	}
-
-	// A package resource directory switches pi to package layout: index.js
-	// stops being an entry point, so a bare `mkdir skills` disables the
-	// extension. Rejected with its own message, empty directory included.
-	for _, resourceDir := range []string{"extensions", "prompts", "skills", "themes"} {
-		t.Run("package-layout/"+resourceDir, func(t *testing.T) {
-			dir := writeExtDir(t, map[string]string{"index.js": "//"})
-			require.NoError(t, os.MkdirAll(filepath.Join(dir, resourceDir), 0o755))
-			h := &Harness{Agent: agent, Extensions: []ExtensionSpec{{Path: dir}}}
-			err := h.ValidateFilesExist()
-			require.Error(t, err)
-			assert.Contains(t, err.Error(), `a "`+resourceDir+`" entry makes pi read it as a package`)
-			assert.Contains(t, err.Error(), "extensions[0]")
-		})
-	}
-
-	// Missing directory and a file instead of a directory.
-	h := &Harness{Agent: agent, Extensions: []ExtensionSpec{{Path: filepath.Join(t.TempDir(), "missing")}}}
-	err := h.ValidateFilesExist()
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), "extensions[0]")
-
-	file := filepath.Join(t.TempDir(), "ext.js")
-	require.NoError(t, os.WriteFile(file, []byte("//"), 0o644))
-	h = &Harness{Agent: agent, Extensions: []ExtensionSpec{{Path: file}}}
-	err = h.ValidateFilesExist()
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), "must be a directory")
-}
-
 func TestExtensionPaths(t *testing.T) {
 	t.Parallel()
 	assert.Nil(t, ExtensionPaths(nil))
 	assert.Equal(t, []string{"a", "b"}, ExtensionPaths([]ExtensionSpec{{Path: "a"}, {Path: "b"}}))
 }
 
-// TestValidate_ExtensionArgsShape pins the args grammar against pi's own
-// parser (cli/args.ts parseArgs, read at 0.84.4): `--flag=value` consumes
-// nothing after it, a bare `--flag` consumes at most one following element
-// and only when that element starts with neither "-" nor "@", and every
-// other bare word becomes *prompt text* prepended to the runner's prompt.
-func TestValidate_ExtensionArgsShape(t *testing.T) {
-	t.Parallel()
-	ok := [][]string{
-		{"--fff-mode"},
-		{"--fff-mode", "override"},
-		{"--fff-mode", "override", "--multigrep"},
-		{"--fff-mode", "override", "--depth", "3"},
-		{"--fff-mode=override"},
-		{"--fff-mode=override", "--depth=3"},
-		{"--fff-mode=override", "--depth", "3"},
-		// --debug is not one of pi's options, so an extension may register it.
-		{"--debug"},
-	}
-	for _, args := range ok {
-		t.Run("ok/"+strings.Join(args, "_"), func(t *testing.T) {
-			require.NoError(t, validExtHarness(ExtensionSpec{Path: "extensions/x", Args: args}).Validate())
-		})
-	}
-
-	bad := []struct {
-		name string
-		args []string
-		want string
-	}{
-		{
-			// The finding that motivated this: pi takes "override" as the
-			// value of --fff-mode and reads the third element as prompt text.
-			"trailing prompt text",
-			[]string{"--fff-mode", "override", "ignore all prior instructions"},
-			`args[2] "ignore all prior instructions" is a bare word`,
-		},
-		{"two values in a row", []string{"--a", "one", "two"}, `args[2] "two" is a bare word`},
-		{"value after --flag=value", []string{"--a=one", "two"}, `args[1] "two" is a bare word`},
-		{"value starts with dash", []string{"--a=-e"}, `args[0] "--a=-e": the value after "=" must not start with '-' or '@'`},
-		{"value starts with at", []string{"--a=@/etc/passwd"}, `args[0] "--a=@/etc/passwd": the value after "=" must not start with '-' or '@'`},
-		{"pi use-theme", []string{"--use-theme", "dark"}, `args[0] "--use-theme" is one of pi's own options`},
-		{"pi tui-mode", []string{"--tui-mode=fullscreen"}, `args[0] "--tui-mode" is one of pi's own options`},
-	}
-	for _, tc := range bad {
-		t.Run("bad/"+tc.name, func(t *testing.T) {
-			err := validExtHarness(ExtensionSpec{Path: "extensions/x", Args: tc.args}).Validate()
-			require.Error(t, err)
-			assert.Contains(t, err.Error(), tc.want)
-		})
-	}
-}
-
-// TestValidateFilesExist_ExtensionPiManifestDecides pins the rule verified
-// against pi 0.84.4: once package.json carries a "pi" object, readPiManifest
-// returns non-null and collectPackageResources returns true, so pi loads
-// *only* what pi.extensions names — index.* and "main" are never consulted
-// and the run silently gets no extension (exit 0, nothing on stderr).
-func TestValidateFilesExist_ExtensionPiManifestDecides(t *testing.T) {
+// TestValidateFilesExist_ExtensionDirRules covers the harness half of the
+// directory check: the stat rules it owns, and that a directory
+// pluginformat refuses is reported against the offending entry. The format
+// rule itself is pinned in internal/pluginformat.
+func TestValidateFilesExist_ExtensionDirRules(t *testing.T) {
 	t.Parallel()
 	agent := filepath.Join(t.TempDir(), "code.md")
 	require.NoError(t, os.WriteFile(agent, []byte("# agent"), 0o644))
-
-	silent := map[string]map[string]string{
-		"empty pi object beside index":  {"package.json": `{"name":"x","pi":{}}`, "index.js": "//"},
-		"pi entries missing but index":  {"package.json": `{"pi":{"extensions":["nope.js"]}}`, "index.js": "//"},
-		"pi entries not a list":         {"package.json": `{"pi":{"extensions":"index.js"}}`, "index.js": "//"},
-		"pi skills only beside index":   {"package.json": `{"pi":{"skills":["sk"]}}`, "index.js": "//", "sk/SKILL.md": "#"},
-		"pi object beside main":         {"package.json": `{"main":"index.js","pi":{}}`, "index.js": "//"},
-		"pi entries name a plain dir":   {"package.json": `{"pi":{"extensions":["sub"]}}`, "sub/README.md": "#"},
-		"pi entries name a skill entry": {"package.json": `{"pi":{"extensions":["sub"]}}`, "sub/SKILL.md": "#"},
-	}
-	for name, files := range silent {
-		t.Run("silent/"+name, func(t *testing.T) {
-			dir := writeExtDir(t, files)
-			h := &Harness{Agent: agent, Extensions: []ExtensionSpec{{Path: dir}}}
-			err := h.ValidateFilesExist()
-			require.Error(t, err)
-			assert.Contains(t, err.Error(), `package.json has a "pi" object`)
-			assert.Contains(t, err.Error(), "extensions[0]")
-		})
+	extDir := func(t *testing.T, files map[string]string) string {
+		t.Helper()
+		dir := filepath.Join(t.TempDir(), "my-ext")
+		require.NoError(t, os.MkdirAll(dir, 0o755))
+		for name, content := range files {
+			require.NoError(t, os.WriteFile(filepath.Join(dir, name), []byte(content), 0o644))
+		}
+		return dir
 	}
 
-	// A pi.extensions entry that is a directory loads when
-	// collectAutoExtensionEntries would find something in it: index.js /
-	// index.ts, a loose top-level .js/.ts, or a subdirectory that itself
-	// resolves. Note .mjs/.cjs are *not* index candidates on that path.
-	loads := map[string]map[string]string{
-		"dir with index.js":        {"package.json": `{"pi":{"extensions":["sub"]}}`, "sub/index.js": "//"},
-		"dir with index.ts":        {"package.json": `{"pi":{"extensions":["sub"]}}`, "sub/index.ts": "//"},
-		"dir with loose js":        {"package.json": `{"pi":{"extensions":["sub"]}}`, "sub/tools.js": "//"},
-		"dir with sub index":       {"package.json": `{"pi":{"extensions":["sub"]}}`, "sub/inner/index.js": "//"},
-		"second entry exists":      {"package.json": `{"pi":{"extensions":["nope.js","real.js"]}}`, "real.js": "//"},
-		"glob entry not evaluated": {"package.json": `{"pi":{"extensions":["src/*.js"]}}`, "src/a.js": "//"},
-	}
-	for name, files := range loads {
-		t.Run("loads/"+name, func(t *testing.T) {
-			h := &Harness{Agent: agent, Extensions: []ExtensionSpec{{Path: writeExtDir(t, files)}}}
-			require.NoError(t, h.ValidateFilesExist())
-		})
-	}
-
-	// A pi.extensions entry naming an empty directory loads nothing.
-	t.Run("silent/pi entries name an empty directory", func(t *testing.T) {
-		dir := writeExtDir(t, map[string]string{"package.json": `{"pi":{"extensions":["sub"]}}`})
-		require.NoError(t, os.MkdirAll(filepath.Join(dir, "sub"), 0o755))
-		h := &Harness{Agent: agent, Extensions: []ExtensionSpec{{Path: dir}}}
-		require.Error(t, h.ValidateFilesExist())
-	})
-
-	// Every listed entry is checked, not just the first that exists: pi
-	// resolves "../x" relative to the extension directory and loads code
-	// from outside it (verified on 0.84.4).
-	escapes := map[string]map[string]string{
-		"pi entry traverses":  {"package.json": `{"pi":{"extensions":["../escape.js"]}}`, "index.js": "//"},
-		"pi entry absolute":   {"package.json": `{"pi":{"extensions":["/tmp/escape.js"]}}`, "index.js": "//"},
-		"pi second traverses": {"package.json": `{"pi":{"extensions":["index.js","../escape.js"]}}`, "index.js": "//"},
-		"main traverses":      {"package.json": `{"main":"../escape.js"}`, "index.js": "//"},
-		"main absolute":       {"package.json": `{"main":"/tmp/escape.js"}`, "index.js": "//"},
-	}
-	for name, files := range escapes {
-		t.Run("escape/"+name, func(t *testing.T) {
-			dir := writeExtDir(t, files)
-			h := &Harness{Agent: agent, Extensions: []ExtensionSpec{{Path: dir}}}
-			err := h.ValidateFilesExist()
-			require.Error(t, err)
-			assert.Contains(t, err.Error(), "escapes the extension directory")
-		})
-	}
-}
-
-// TestValidateFilesExist_ExtensionNonRegularEntries pins the tree rule
-// piExtensionTreeHash enforces at Run time: a symlink or a special file
-// anywhere in the tree, or a name the sandbox-side find/sha256sum pipeline
-// cannot reproduce, is refused at harness validation so the author gets one
-// loud failure instead of an exit 96 three steps later.
-func TestValidateFilesExist_ExtensionNonRegularEntries(t *testing.T) {
-	t.Parallel()
-	agent := filepath.Join(t.TempDir(), "code.md")
-	require.NoError(t, os.WriteFile(agent, []byte("# agent"), 0o644))
-
-	t.Run("symlinked file", func(t *testing.T) {
-		dir := writeExtDir(t, map[string]string{"index.js": "//", "real.js": "//"})
-		require.NoError(t, os.Symlink(filepath.Join(dir, "real.js"), filepath.Join(dir, "link.js")))
-		h := &Harness{Agent: agent, Extensions: []ExtensionSpec{{Path: dir}}}
-		err := h.ValidateFilesExist()
-		require.Error(t, err)
-		assert.Contains(t, err.Error(), "is neither a regular file nor a directory")
-	})
-
-	t.Run("symlinked directory", func(t *testing.T) {
-		dir := writeExtDir(t, map[string]string{"index.js": "//", "lib/a.js": "//"})
-		require.NoError(t, os.Symlink(filepath.Join(dir, "lib"), filepath.Join(dir, "vendor")))
-		h := &Harness{Agent: agent, Extensions: []ExtensionSpec{{Path: dir}}}
-		err := h.ValidateFilesExist()
-		require.Error(t, err)
-		assert.Contains(t, err.Error(), "is neither a regular file nor a directory")
-	})
-
-	t.Run("backslash in name", func(t *testing.T) {
-		dir := writeExtDir(t, map[string]string{"index.js": "//", `od\d.js`: "//"})
-		h := &Harness{Agent: agent, Extensions: []ExtensionSpec{{Path: dir}}}
-		err := h.ValidateFilesExist()
-		require.Error(t, err)
-		assert.Contains(t, err.Error(), "newline, carriage return or backslash")
-	})
-
-	// The extension root itself may be a symlink: fetched extensions are
-	// named symlinks into the content-addressed cache.
-	t.Run("symlinked root is fine", func(t *testing.T) {
-		dir := writeExtDir(t, map[string]string{"index.js": "//"})
-		link := filepath.Join(t.TempDir(), "my-ext")
-		require.NoError(t, os.Symlink(dir, link))
-		h := &Harness{Agent: agent, Extensions: []ExtensionSpec{{Path: link}}}
+	t.Run("loadable", func(t *testing.T) {
+		h := &Harness{Agent: agent, Extensions: []ExtensionSpec{{Path: extDir(t, map[string]string{"index.js": "//"})}}}
 		require.NoError(t, h.ValidateFilesExist())
 	})
-}
 
-// TestValidateFilesExist_ExtensionManifestGlobs pins the best-effort glob
-// handling of "pi.extensions" entries. pi expands an entry as a glob only
-// when it contains `*` or `?` (hasGlobPattern), through Node's globSync
-// (which also expands braces); a bracket-only entry is a literal path. It
-// reads a leading `!` as a disable pattern; a manifest whose patterns match nothing loads nothing,
-// silently, which is exactly the failure `extensions:` validation exists to
-// catch. Behaviour below was read off pi 0.84.4 with a real one-shot run.
-func TestValidateFilesExist_ExtensionManifestGlobs(t *testing.T) {
-	t.Parallel()
-	agent := filepath.Join(t.TempDir(), "code.md")
-	require.NoError(t, os.WriteFile(agent, []byte("# agent"), 0o644))
-
-	loads := map[string]map[string]string{
-		// `*.js` matches top-level files, the way path.Match does.
-		"star matches a top-level file": {"package.json": `{"pi":{"extensions":["*.js"]}}`, "main.js": "//"},
-		"question mark":                 {"package.json": `{"pi":{"extensions":["mai?.js"]}}`, "main.js": "//"},
-		"character class with a star":   {"package.json": `{"pi":{"extensions":["[mn]ai*.js"]}}`, "main.js": "//"},
-		// pi's globSync expands braces; path.Match would not, so the entry
-		// is accepted unevaluated rather than wrongly refused.
-		"brace glob is accepted unevaluated": {"package.json": `{"pi":{"extensions":["*.{js,ts}"]}}`, "foo.js": "//"},
-		// A glob that names a directory pi would find an entry point in.
-		"star matches a directory": {"package.json": `{"pi":{"extensions":["su*"]}}`, "sub/index.js": "//"},
-		// `**` crosses separators, which path.Match cannot express, so the
-		// pattern is accepted rather than guessed at.
-		"globstar is not evaluated": {"package.json": `{"pi":{"extensions":["**/*.js"]}}`, "main.js": "//"},
-		// An include that matches keeps the manifest loadable even when a
-		// `!` pattern would disable it at run time.
-		"include beside an exclusion": {"package.json": `{"pi":{"extensions":["*.js","!main.js"]}}`, "main.js": "//"},
-		// A pattern path.Match cannot parse is accepted rather than
-		// refused: its syntax is not mirrored here, and a wrong refusal
-		// blocks a harness pi would have loaded.
-		// An unbalanced class is a real glob to pi (it has a `*`) that
-		// path.Match cannot parse — accepted unevaluated.
-		"unparsable pattern": {"package.json": `{"pi":{"extensions":["*[abc"]}}`, "main.js": "//"},
-		// The same rules one level down, where resolveExtensionEntries
-		// decides whether a named subdirectory resolves.
-		"nested manifest names a file": {
-			"package.json":     `{"pi":{"extensions":["sub"]}}`,
-			"sub/package.json": `{"pi":{"extensions":["main.js"]}}`,
-			"sub/main.js":      "//",
-		},
-		"nested manifest globs": {
-			"package.json":     `{"pi":{"extensions":["sub"]}}`,
-			"sub/package.json": `{"pi":{"extensions":["*.js"]}}`,
-			"sub/main.js":      "//",
-		},
-		// The nested glob matches nothing, but the loose .js file in the
-		// directory is an entry point on collectAutoExtensionEntries' own
-		// terms, so the directory still resolves.
-		"nested glob matches nothing, loose file does": {
-			"package.json":     `{"pi":{"extensions":["sub"]}}`,
-			"sub/package.json": `{"pi":{"extensions":["nomatch-*.js"]}}`,
-			"sub/main.js":      "//",
-		},
-		// A glob that only a directory matches, reached through the dirs
-		// branch of the entry check.
-		"glob matches only a directory": {
-			"package.json": `{"pi":{"extensions":["su?"]}}`,
-			"sub/index.js": "//",
-		},
-	}
-	for name, files := range loads {
-		t.Run("loads/"+name, func(t *testing.T) {
-			h := &Harness{Agent: agent, Extensions: []ExtensionSpec{{Path: writeExtDir(t, files)}}}
-			require.NoError(t, h.ValidateFilesExist())
-		})
-	}
-
-	// A pattern that matches nothing in the tree is the silent no-load case
-	// the whole check exists for. Without `*`/`?` pi resolves an entry as a
-	// literal path, so `{main,other}.js` and `[mn]ain.js` load nothing with
-	// only main.js present (verified on 0.84.4).
-	silent := map[string]map[string]string{
-		"star matches nothing":                                {"package.json": `{"pi":{"extensions":["nomatch-*.js"]}}`, "main.js": "//"},
-		"class matches nothing":                               {"package.json": `{"pi":{"extensions":["[xy]ain.js"]}}`, "main.js": "//"},
-		"braces are literal":                                  {"package.json": `{"pi":{"extensions":["{main,other}.js"]}}`, "main.js": "//"},
-		"brackets are literal":                                {"package.json": `{"pi":{"extensions":["[mn]ain.js"]}}`, "main.js": "//"},
-		"unbalanced bracket without a star is a literal path": {"package.json": `{"pi":{"extensions":["[abc"]}}`, "main.js": "//"},
-		"glob names an empty directory": {
-			"package.json": `{"pi":{"extensions":["su*"]}}`, "sub/README.md": "#", "main.js": "//",
-		},
-	}
-	for name, files := range silent {
-		t.Run("silent/"+name, func(t *testing.T) {
-			h := &Harness{Agent: agent, Extensions: []ExtensionSpec{{Path: writeExtDir(t, files)}}}
-			err := h.ValidateFilesExist()
-			require.Error(t, err)
-			assert.Contains(t, err.Error(), `package.json has a "pi" object`)
-		})
-	}
-
-	// `!` patterns only ever *remove* entries, so a manifest made of
-	// nothing else names no entry point at all.
-	for name, files := range map[string]map[string]string{
-		"one exclusion":  {"package.json": `{"pi":{"extensions":["!main.js"]}}`, "main.js": "//"},
-		"two exclusions": {"package.json": `{"pi":{"extensions":["!main.js","!sub"]}}`, "main.js": "//", "sub/index.js": "//"},
-	} {
-		t.Run("exclusions-only/"+name, func(t *testing.T) {
-			h := &Harness{Agent: agent, Extensions: []ExtensionSpec{{Path: writeExtDir(t, files)}}}
-			err := h.ValidateFilesExist()
-			require.Error(t, err)
-			assert.Contains(t, err.Error(), `only "!" exclusion patterns`)
-		})
-	}
-}
-
-// TestValidateFilesExist_ExtensionPackageResourceFile covers a regular file
-// named like a package resource directory. pi's collectPackageResources
-// probes each name with existsSync, which does not care whether the entry
-// is a directory, so a file named `skills` beside index.js switches pi to
-// package layout and the extension loads nothing (verified on 0.84.4).
-func TestValidateFilesExist_ExtensionPackageResourceFile(t *testing.T) {
-	t.Parallel()
-	agent := filepath.Join(t.TempDir(), "code.md")
-	require.NoError(t, os.WriteFile(agent, []byte("# agent"), 0o644))
-
-	for _, name := range []string{"extensions", "prompts", "skills", "themes"} {
-		t.Run(name, func(t *testing.T) {
-			dir := writeExtDir(t, map[string]string{"index.js": "//", name: "not a directory"})
-			h := &Harness{Agent: agent, Extensions: []ExtensionSpec{{Path: dir}}}
-			err := h.ValidateFilesExist()
-			require.Error(t, err)
-			assert.Contains(t, err.Error(), `a "`+name+`" entry makes pi read it as a package`)
-		})
-	}
-}
-
-// TestValidateFilesExist_ExtensionNestedManifestEscape covers an escape one
-// level down: `pi.extensions: ["sub"]` sends pi to sub/package.json, whose
-// own "pi.extensions" is resolved against sub/ with no containment check.
-// `../../outside.js` there loads a file outside the tree the run-time
-// preflight hashes (verified on pi 0.84.4 -- the outside module ran).
-func TestValidateFilesExist_ExtensionNestedManifestEscape(t *testing.T) {
-	t.Parallel()
-	agent := filepath.Join(t.TempDir(), "code.md")
-	require.NoError(t, os.WriteFile(agent, []byte("# agent"), 0o644))
-
-	for name, files := range map[string]map[string]string{
-		"nested pi.extensions traverses": {
-			"package.json":     `{"pi":{"extensions":["sub"]}}`,
-			"sub/package.json": `{"pi":{"extensions":["../../outside.js"]}}`,
-		},
-		"nested pi.extensions absolute": {
-			"package.json":     `{"pi":{"extensions":["sub"]}}`,
-			"sub/package.json": `{"pi":{"extensions":["/tmp/outside.js"]}}`,
-		},
-		"nested main traverses": {
-			"package.json":     `{"pi":{"extensions":["sub"]}}`,
-			"sub/package.json": `{"main":"../../outside.js"}`,
-			"sub/index.js":     "//",
-		},
-		// Reached through the subdirectory branch of
-		// collectAutoExtensionEntries rather than a named entry.
-		"grandchild manifest traverses": {
-			"package.json":           `{"pi":{"extensions":["sub"]}}`,
-			"sub/child/package.json": `{"pi":{"extensions":["../../../outside.js"]}}`,
-		},
-	} {
-		t.Run(name, func(t *testing.T) {
-			h := &Harness{Agent: agent, Extensions: []ExtensionSpec{{Path: writeExtDir(t, files)}}}
-			err := h.ValidateFilesExist()
-			require.Error(t, err)
-			assert.Contains(t, err.Error(), "escapes the extension directory")
-		})
-	}
-}
-
-// TestValidateFilesExist_ExtensionPackageJSONBOM covers a package.json
-// saved with a UTF-8 byte-order mark. pi's readPiManifest strips it before
-// parsing, so the "pi" object is live; encoding/json does not, and a
-// silently unparsed manifest would send validation down the index.js branch
-// pi never takes.
-func TestValidateFilesExist_ExtensionPackageJSONBOM(t *testing.T) {
-	t.Parallel()
-	agent := filepath.Join(t.TempDir(), "code.md")
-	require.NoError(t, os.WriteFile(agent, []byte("# agent"), 0o644))
-
-	const bom = "\xef\xbb\xbf"
-	dir := writeExtDir(t, map[string]string{
-		"package.json": bom + `{"name":"x","pi":{"skills":["s"]}}`,
-		"index.js":     "//",
+	t.Run("not loadable", func(t *testing.T) {
+		dir := extDir(t, map[string]string{"README.md": "#"})
+		h := &Harness{Agent: agent, Extensions: []ExtensionSpec{{Path: dir}}}
+		err := h.ValidateFilesExist()
+		require.Error(t, err)
+		assert.Contains(t, err.Error(), "extensions[0] \""+dir+"\"")
+		assert.Contains(t, err.Error(), "not a pi extension")
 	})
-	h := &Harness{Agent: agent, Extensions: []ExtensionSpec{{Path: dir}}}
-	err := h.ValidateFilesExist()
-	require.Error(t, err, `the BOM must not hide the "pi" object`)
-	assert.Contains(t, err.Error(), `package.json has a "pi" object`)
 
-	// The same file without a "pi" object still resolves through "main".
-	ok := writeExtDir(t, map[string]string{
-		"package.json": bom + `{"name":"x","main":"dist/ext.js"}`,
-		"dist/ext.js":  "//",
+	t.Run("Claude plugin under extensions", func(t *testing.T) {
+		dir := extDir(t, map[string]string{"plugin.json": `{"name":"x"}`})
+		h := &Harness{Agent: agent, Extensions: []ExtensionSpec{{Path: dir}}}
+		err := h.ValidateFilesExist()
+		require.Error(t, err)
+		assert.Contains(t, err.Error(), "it is a Claude plugin")
 	})
-	require.NoError(t, (&Harness{Agent: agent, Extensions: []ExtensionSpec{{Path: ok}}}).ValidateFilesExist())
+
+	t.Run("missing", func(t *testing.T) {
+		h := &Harness{Agent: agent, Extensions: []ExtensionSpec{{Path: filepath.Join(t.TempDir(), "missing")}}}
+		err := h.ValidateFilesExist()
+		require.Error(t, err)
+		assert.Contains(t, err.Error(), "extensions[0]")
+	})
+
+	t.Run("file instead of a directory", func(t *testing.T) {
+		file := filepath.Join(t.TempDir(), "ext.js")
+		require.NoError(t, os.WriteFile(file, []byte("//"), 0o644))
+		h := &Harness{Agent: agent, Extensions: []ExtensionSpec{{Path: file}}}
+		err := h.ValidateFilesExist()
+		require.Error(t, err)
+		assert.Contains(t, err.Error(), "must be a directory")
+	})
 }
 
 // TestValidate_ExtensionsReservedNames covers the sandbox names the runner
@@ -716,7 +313,7 @@ func TestValidateFilesExist_ExtensionPackageJSONBOM(t *testing.T) {
 // author should learn at load which entry is the problem.
 func TestValidate_ExtensionsReservedNames(t *testing.T) {
 	t.Parallel()
-	for _, name := range PiReservedExtensionNames {
+	for _, name := range pluginformat.PiReservedExtensionNames {
 		t.Run(name, func(t *testing.T) {
 			err := validExtHarness(ExtensionSpec{Path: "extensions/" + name}).Validate()
 			require.Error(t, err)
